requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-image>=0.20.0
imageio-ffmpeg>=0.4.0
customtkinter>=5.2.0
python-dotenv>=1.0.0
requests>=2.31.0